Added inline CSS directly in our JSX
We have to refactor the implicit return to explicit and then provided a js object directly inside our style object and so we have two {}s, one inside the other.  These are inline styles. In JS object if we have CSS property that has hyphen, we will have to capitalise and the values will have to be strings. So the general CSS rules do not apply if we have both the inline and normal CSS properties outside of the component at the same time.
Sometimes external libraries (eg. a slider library or a modal library ) use inline styles. So we have to figure out how to modify those inline styles to be successful.