Add inline CSS object indirectly
We can create the object inside (that is what we did here) or outside of the component and then we provide the reference inside the return. Notice the values are the same as the objects we had when we had the inline object directly added. The only difference now is instead of passing the object directly we just set up the reference.
Inline styles are good to know but we will stick to external stylesheet, some CSS libraries because adding styles one by one to each element is somewhat painful.