Props and Object Destructuring- 1st way
Instead of props.propsName, we go with props and since it is an object we can destructure the props inside the component  or our function body. We just need to type this:

const { img, title, author } = props;

Plus reference the properties (img, title, author) ones and we are good to go.