Children Prop
Children allows us to access everything we are rendering between the component tags.
This task is just to showcase the general concept of children prop.
We'll only use it when we need to set up context API.

We decided to have a para and a button in one of the components.

Add added some CSS styling.
- added margin to paragraph
- added to "align-items: start;" to make the cards start at one point and end where the content ends