Proper List
We create an array of objects called books.

In react we cannot render or display an object directly like below in JSX:

const BookList = () => {
  return <section className="booklist">{books}</section>;
};

Map: A map method creates a new array from calling a function for every array element.

The solution is to use map method to iterate over the books array data and for every item we will return a component. We pull the properties out of the object, and pass them one by one to the book component. Thus, for every item, we return this component: <Book title= {title} img={img} author={author} />.

And we get this awesome setup.

All our logic is inside the Book component to render stuff on the screen. We simply need to iterate over and for every item we just need to return the component.

We are iterating over the list and we are wrapping the elements in some kind of html structure.