Key Prop
Key has to be a unique value that we use so that React can keep tract of the items in the list it renders. So we  provide a unique value and set it equal to a key prop.

We place the key where we are iterating over the data and then we set on the item that we are returning. It has to be in the element that we are returning when we are iterating over the list.

If the setup is like below, then we add the key on a div and not h2:

function BookList() {
  return (
    <section className='booklist'>
      {books.map((book) => {
        console.log(book);

        // return 'hello';
        return (
          <div>
            <h2>{book.title}</h2>
          </div>
        );
      })}
    </section>
  );
}

Typically, key is the id, but it can be any unique value. It can be a string instead of a number. It is not advised to use index as key if the list is changing.

If we pass the index as key then we can also write as below:
function BookList() {
  return (
    <section className='booklist'>
      {books.map((book, index) => {
        console.log(book);
        const { img, title, author, id } = book;
        return <Book book={book} key={index} />;
      })}
    </section>
  );
}

But if our list is going to change or if we remove items from the list, then the above way will not work and we will get bugs down the road