Event Basics
Rendering a list with the help of components is awesome, but it would be nicer if the user can interact with our application. Otherwise, our app is boring.
Our app needs to respond to events. For eg. the user clicks a button.
We learn how to handle events in react by logging stuff without hooks to control the state.

In vanilla JS we have the following setup (we have some kind of button for input, then we select that input/element and add event listener, where we pass the name and the callback function):

const btn = document.getElementById('btn');

btn.addEventListener('click', function (e) {
  // access event object (the element we are selecting, the value that is inside of it)
  // do something when event fires
});

In react, the setup is extremely similar. We have an element (our element, button is located in the JSX in the component), an event (we "onClick" in Camel case instead of "click" in JS) and a call back function. It is a functionality we invoke when that event fires. This is shown below:

const EventExamples = () => {
  const handleButtonClick = () => {
    alert('handle button click');
  };
  return (
    <section>
      <button onClick={handleButtonClick}>click me</button>
    </section>
  );
};

On here --> <button onClick={handleButtonClick}>click me</button>
We are basically passing in the reference . When we go back to JS land, we set up the curlies and same as how we reference the variable, now we reference the function.

In all our samples we will invoke alert or log because we haven't covered how we can control what we can see on a screen using state.

[React Events](https://reactjs.org/docs/events.html)
- no need to memorize them(idea is the same)
- most common
  - onClick (click events)
  - onSubmit (submit form )
  - onChange (input change )

First we focus on the JSX we are returning. We go to index.js and right after the booklist (which we will eventually remove), we create a component called EventExamples.

Once we covered the general setup, then we create EventExamples component, which we will render above the books one because eventually we will have the books and EventExamples will stay for our reference.  In the EventExamples we set up a section form with h2, input and a button with the onClick.

First we set up the basic return and not worry about the function or the event and only worry about the JSX we are returning. Then we go to the BookList and reference EventExamples.
It's nice to set up the component first as we know that it's definitely being rendered on the screen. There is no point to set up the entire logic and then start chasing some bugs if the initial set up was wrong.

Once the component setup is correct, then we can start adding the logic inside the return inside EventExamples.

We add h2, text input and add inline styles to it,

We have events like onClick (if we want to respond to click event) or onChange (we want to respond to user typing something in the input, the event name is onChange) and then we need to pass in the reference.

We need to set up the functions like handleFormInput or handleButtonClick. They can be inside of the component or outside component, they can be coming from another file - that part is irrelevant.
We need to pass those functions  as a reference and then every time that event will fire, React will execute the functionality we have inside of the reference function.

We notice in the console that  we have logs. Thus, we are responding to the events.  As a result our application is more dynamic.

For events we needed 3 things:
an element, an event and lastly we need reference to a function  or a callback function that we will invoke once the event takes place or fires off.