Form Submission without button
In vanilla JS when we pass the callback function, we have access to an event object. In that object we can find a bunch of useful info and we can do the same thing in React.

So every call back function we are referencing will have access to that event. Since it is a parameter, we can give it whatever name we want eg. e or event.
Let's log it first. We will get a bunch of useful info.
The one we will use the most is event.target.name
Target is going to point back to an element (input in our case). If we want to access the name, we will do event.target.name. If we want the value, we will do event.target.value
This will allow us to collect the data from the input.

We will also use prevent default, which is very useful when we have the form.

If we want to respond to form submission, we will need to set up an event. We will create a function called handleFormSubmission and let's pass in the handle with onSubmit.  Now it doesn't matter if we press enter after typing something in the input or not. If we press enter, we should see something in the browser. The reason for that is because the default behavior of the forms is to collect the values and send it to some URL. Since we don't want to do that and rather handle the forms ourselves, that's why we want to access the event again, same as with handleFormInput, we have right away access to it, and there is a function called event.preventDefault. This means that we will handle those form submissions ourselves and thus we can see the log in the console (if we still can't see it, then we mouse click on the input and then press "enter" on keyboard. So each and every reference function will have access to event. In there we can find useful things and mostly we will work with target, target.name, target.value as well as event.preventDefault.

Next task we will add a button to submit the form which will be our setup for the rest of the course.