Form Submission - Button example Part 1
There are 2 ways to set up a button to handle form submissions in HTML:

- If we set up a button with type submit in the form, we can utilise that button. Keep in mind that the value of the type has to be submit:  type="submit" (if type="button" it will not work)

To showcase this we have two buttons- one with type="submit" and the other with type="button".
If we click the button with type="submit", we nicely submitted the form because we have onSubmit on the form: <form onSubmit={handleFormSubmission}>... and we have the callback function.

If we click the one with type="button", we have the alert.  It will not submit the form.