Form Submission - Button example Part 2
Here is another way if we don't like the setup  as in the previous commit <form onSumbit={handleFormSubmission}>... in the form, we can remove the onSubmit next to the form and write onSubmit next to the button and pass in the value handleFormSubmission. If we now click the submit button we will get an error. In order to properly submit this form, we will need to have onClick because it is a button and then both of them will work. We can also submit the form with enter or click the button.

Thus, if we want to use the button, go with onClick. If we want to use the onSubmit on the form, we need to make sure the button has a type="submit" because type="button" will not work.