Anonymous Function (arrow)
Previously we saw the basic setup of addEventListener in Vanilla JS:

const btn = document.getElementById('btn');

btn.addEventListener('click', function (e) {
  // access event object
  // do something when event fires
});

The second parameter above is an anonymous callback function.
Instead of that we could also set up a variable and assign the function to this variable and pass as the second parameter. Or we could directly pass an arrow function as the second parameter which by default is anonymous.

We can also directly pass an arrow function in React. This is an alternative approach where instead of setting up those functions in the same file or a separate file, we can pass the anonymous function directly. Here is the anonymous function:
() => console.log('hello there')

So we are passing here the arrow function. We are not invoking. We just pass as shown below:

<button onClick={() => console.log('hello there')}>click me</button>

This logic is going to run when we click a button in this case.

We won't submit button this time; essentially it works the same way.
Instead of using reference we will use the arrow function this time and whatever logic we want to put inside. We will have:

One of the major gotchas here is that we still have access to the event object. And if we log:  {(e) => console.log(e.target.value)} , we will still access whatever is in the input every time the user types in the input. These things will not change, but we will see this shorter syntax. There is no rule. We can either use this one or we can also set up the reference.
Mostly it can depend on the logic we have.  If we have this input and we are just accessing the value, often we will see this approach. If we have button click and then there is 20 lines of code, most realistically then we will set up a reference function. Both of them are the same. But in this case we have less lines of code using arrow function.