Components Feature
Components are independent by default.
We can start by removing the events. Events is one of the main features to any application if it's dynamic and we will return to it later.
First we remove the reference of the EventExamples in the BookList component and then remove the EventExamples component.

Let's set up an event in the book component. This is quite straight forward.
We set up a function. We call this display title. We could also set up an anonymous callback function if we wanted. In our case, we have a reference.  We simply want to log a title prop. And we also need to set up the event  and we need to reference the function. It makes sense to have a value inside of the button.

Test: When we click the button we see in the console, the same title that is displayed in a component and this happens for the second component referencing the title that is associated with the second component.

In vanilla JS this will require a lot of work to have this functionality to directly point to that one specific item. In vanilla JS this does not come by default. We can't just right away go to a specific element and say that the functionality will apply to only that element. We need to do more work.

In case of React, whatever functionality will have inside of the component, it is narrowed down to that component. Whereas in vanilla JS in order to have this functionality, we need to jump through the hoops. In React it comes by default.