Prop Drilling
In React, as far as data flow, we can only pass the data down. In our case, we can only pass data from BookList to a Book component, not the other way round.

If we have a component nested inside of the Book component, if we want to pass something from the BookList, all the way down to that component, we will have to go through the Book component.

There are alternatives. We will cover, API, Redux, we can use other state libraries for that as well, but by default, we can only pass the props down.

Essentially this process is called "prop drilling".

We will not need the title, but we will keep the button to implement it in this example.

Inside the BookList we create a variable named someValue and set up a function named displayValue, where we are accessing the variable that is in the BookList component.

How do we pass this displayValue function down?
It is effectively the same thing, we need to come up with a prop name and we need to pass it as a value.  We are iterating over the book and we have this Book component where we will pass our displayValue function as props like we did before with other components in JSX.

In order to access the displayValue in the Book component, as it is already passed as a props in the parent component, we will add it as the props in the destructured object and pass it as our reference in the onClick.

In our dev tools we will see we nicely log shakeAndBake.

The main points are:
- We can only pass props down from the parent to the child.
- It means that we have really nested structure, we have to go through all of those components. The passing of props literally down to a grandchild (if there was a child component inside of the book components, then we will do the same thing).

Essentially we don't use it in the Book component, but we pass it down to the other component in JSX, it becomes prop drilling.