Update index.js
To cover a more complex example...We have a function in the component and we need to invoke it in the child.
We start with a little challenge where in the BookList component, we want to create a GetBook function. This function accepts ideas and arguments, and it finds a book.

We will use the find method.
We will grab the books array and we find the book that matches the ID that we are passing in. And we will log our result.

We just set up the function, log the book that matches the ID that we passed in. Then pass that function down to the Book component and try to invoke it on a button.

Then we will hit a bug and that is the purpose of this task.