ES6 modules - export/import with the named one
We split up our application into multiple files, folders and the ability to import and export the code which helps us to organise our application. As our  application grows it is not sustainable to have such big index.js file. So we split up the code where we have separate functionalities in different places. If we need the component or data, we just access it. We use ES6 modules to do that. This is not unique to React.

When it comes to export and import, we have 2 flavours- a named one and a default one. If we have one component in the file,  we do that as a default one. If we have multiples, they will be named.

When we import with a named one, we need to use the same name. With the default one, we can come up with whatever we want.

So in the src we set up two files- books.js forour data and Book.js for our component. We remove books array from index.js and add it to Books.js

When we import books.js inside index.js we use curlies which means we will get specific item i.e. the item we are exporting is exported as named export.  Sp {book} needs to match what is exactly in the file.