ES6 modules - export/import with the default flavour
We have split up the code from index.js to its component and data file.

In this default flavour, when importing our component in index.js we don't have to put curlies and we are importing anything in default, we can name it however name we want in index.js (banana in our example).

Here, we are doing all the Books related work in Books.js So we won't have to run back to index.js As long as the import is correct, we are good to go.

These ES6 modules are not specific to React. We can use them in vanilla JS, in Node to help us to organise our application.