Local Images (src folder)
Setting up local images in the src folder is a better option for setting up the assets because when the React builds our production application, it optimises those assets. Unlike the public one there is no optimisation, we just dump the asset and that's it.
In this case, React is trying to help us by optimising those assets. The result is our application will be faster.

We added one more book into our array. We will utilise the ES6 modules we've already covered.

We have added new images. The folder structure is up to us. For bigger projects, we will have assets and then images folder and CSS folder.
For now, we create images and place it in the source, but it's a must to be in the source.

Even if the structure is more nested, we need to make sure that the parent, the main directory is still within the source. It will not work if it's in the public folder. Now we want to import those images one by one. This is the downside. If we have 100 images we have to do this one by one.

If we have images in the public folder, then we can just reference it as ./ then the path.
This is a bit annoying technically as we need to do it one by one, but the results will be better. Essentially, we want to import and we want to come up with a name and we need to look for images and then the file name and then give the full path extension as below:

 import img1 from "./images/book-1.jpg";

And then instead of the local one or the external image link we simply pass in that variable as below:

img: img1,

We are no longer working in index.js. We only focus on books.js and grab the images one by one.

Once we save, we should have our list of 3 books and the images work as expected.
The main benefit of such approach is that for all the assets including images, they get optimised when React or create-react-app (the tool we are using) builds out our production already ready application. As a result our applications are faster.